+++ pillow_heif/private.py
@@ -6,6 +6,8 @@ from math import ceil
 from typing import Union
 
 from _pillow_heif_cffi import ffi, lib
+from packaging.version import parse as parse_version
+from PIL import __version__ as pil_version
 
 from ._libheif_ctx import LibHeifCtxWrite
 from .constants import HeifChroma, HeifColorProfileType, HeifColorspace
@@ -200,6 +202,15 @@ def set_metadata(ctx: LibHeifCtxWrite, heif_img_handle, info: dict) -> None:
             metadata["content_type"],
         )
         check_libheif_error(error)
+
+
+def exif_from_pillow(additional_info: dict, frame) -> None:
+    if "exif" not in additional_info:
+        if hasattr(frame, "getexif"):
+            if parse_version(pil_version) >= parse_version("9.2.0"):
+                exif = frame.getexif()
+                if exif:
+                    additional_info["exif"] = exif.tobytes()
 
 
 def xmp_from_pillow(additional_info: dict, frame) -> None: